Use mingw-w64 for the wine libraries build
Summary:
According to the [official release notes of version 5.0](https://www.winehq.org/announce/5.0) of Wine this change helps with copy protection schemes that check the on-disk and in-memory contents of the modules.

Building with mingw-w64 also brings up an advantage over the `.dll.so` shims regarding installation sizes, 248.6 MB to 161.7 MB for Wine64 and 235.0 MB to 152.0 MB for Wine32.

**Packaging changes**:
- Enabled kerberos support.
- Enabled gstreamer support for multimedia decoding and encoding.

Depends on D11699.

Signed-off-by: Martin Reboredo <yakoyoku@gmail.com>

Test Plan: Run a D3D9 game using DXVK with both `wine64` and `wine`.

Reviewers: #triage_team, JoshStrobl

Reviewed By: #triage_team, JoshStrobl

Subscribers: JoshStrobl

Differential Revision: https://dev.getsol.us/D11701